+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [0.4.0] - 2021-01-03
 ### Added
 - `Arguments::opt_free_from_*`.
 
@@ -65,7 +67,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - Do not panic while parsing non UTF-8 arguments.
 
-[Unreleased]: https://github.com/RazrFalcon/pico-args/compare/v0.3.4...HEAD
+[Unreleased]: https://github.com/RazrFalcon/pico-args/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/RazrFalcon/pico-args/compare/v0.3.4...v0.4.0
 [0.3.4]: https://github.com/RazrFalcon/pico-args/compare/v0.3.3...v0.3.4
 [0.3.3]: https://github.com/RazrFalcon/pico-args/compare/v0.3.2...v0.3.3
 [0.3.2]: https://github.com/RazrFalcon/pico-args/compare/v0.3.1...v0.3.2